requirements.txt
django
djangorestframework
django-filter
djangorestframework-simplejwt
psycopg2-binary
celery
redis
django-celery-beat
requests
stripe
drf-yasg
django-environ
coverage
ipython
pillow

